Add close-chromium-download-bar.
This script is one of the most (1/2)*(lazy + anti-lazy) things
in my ~/bin. Meaning, I'm not sure if it counts as lazy or the
opposite to write a script that lets us automate clicking a mouse
button using a keyboard combination, but either way, stuff like this
makes me smile.